+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Controllers/ActividadpwaJpaController.java
@@ -18,13 +18,12 @@ import ResPwAEntities.ActividadPwa;
 import ResPwAEntities.Controllers.exceptions.IllegalOrphanException;
 import ResPwAEntities.Controllers.exceptions.NonexistentEntityException;
 import ResPwAEntities.Controllers.exceptions.PreexistingEntityException;
-import java.math.BigDecimal;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 
 /**
  *
- * @author USER
+ * @author tesispepper
  */
 public class ActividadPwaJpaController implements Serializable {
 
@@ -163,7 +162,7 @@ public class ActividadPwaJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                BigDecimal id = actividadPwa.getId();
+                Integer id = actividadPwa.getId();
                 if (findActividadPwa(id) == null) {
                     throw new NonexistentEntityException("The actividadPwa with id " + id + " no longer exists.");
                 }
@@ -176,7 +175,7 @@ public class ActividadPwaJpaController implements Serializable {
         }
     }
 
-    public void destroy(BigDecimal id) throws IllegalOrphanException, NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -239,7 +238,7 @@ public class ActividadPwaJpaController implements Serializable {
         }
     }
 
-    public ActividadPwa findActividadPwa(BigDecimal id) {
+    public ActividadPwa findActividadPwa(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(ActividadPwa.class, id);
