+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Controllers/AntecedenteJpaController.java
@@ -14,7 +14,6 @@ import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import ResPwAEntities.Regla;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -22,7 +21,7 @@ import javax.persistence.EntityManagerFactory;
 
 /**
  *
- * @author USER
+ * @author tesispepper
  */
 public class AntecedenteJpaController implements Serializable {
 
@@ -99,7 +98,7 @@ public class AntecedenteJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                BigDecimal id = antecedente.getId();
+                Integer id = antecedente.getId();
                 if (findAntecedente(id) == null) {
                     throw new NonexistentEntityException("The antecedente with id " + id + " no longer exists.");
                 }
@@ -112,7 +111,7 @@ public class AntecedenteJpaController implements Serializable {
         }
     }
 
-    public void destroy(BigDecimal id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -162,7 +161,7 @@ public class AntecedenteJpaController implements Serializable {
         }
     }
 
-    public Antecedente findAntecedente(BigDecimal id) {
+    public Antecedente findAntecedente(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Antecedente.class, id);
