+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Controllers/RobotJpaController.java
@@ -15,7 +15,6 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import ResPwAEntities.Emocion;
 import ResPwAEntities.Robot;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -23,7 +22,7 @@ import javax.persistence.EntityManagerFactory;
 
 /**
  *
- * @author USER
+ * @author tesispepper
  */
 public class RobotJpaController implements Serializable {
 
@@ -116,7 +115,7 @@ public class RobotJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                BigDecimal id = robot.getId();
+                Integer id = robot.getId();
                 if (findRobot(id) == null) {
                     throw new NonexistentEntityException("The robot with id " + id + " no longer exists.");
                 }
@@ -129,7 +128,7 @@ public class RobotJpaController implements Serializable {
         }
     }
 
-    public void destroy(BigDecimal id) throws IllegalOrphanException, NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -185,7 +184,7 @@ public class RobotJpaController implements Serializable {
         }
     }
 
-    public Robot findRobot(BigDecimal id) {
+    public Robot findRobot(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Robot.class, id);
